+++ packages/core/docs/commercetools/changelog.md
@@ -13,6 +13,7 @@
 - implemented useReviews on product page ([#4800](https://github.com/DivanteLtd/vue-storefront/issues/4800))
 - implemented faceting using useFacet factory ([#4853](https://github.com/DivanteLtd/vue-storefront/issues/4853))
 - fixed anonymous token loading ([#4917](https://github.com/DivanteLtd/vue-storefront/issues/4917))
+- fixed bugs related to customQuery ([#4933](https://github.com/DivanteLtd/vue-storefront/issues/4933), [#4913](https://github.com/DivanteLtd/vue-storefront/issues/4913))
 
 ## 0.1.0